+++ ocdsextensionregistry/extension_version.py
@@ -3,16 +3,15 @@ import json
 import os
 import re
 from contextlib import closing
-from io import BytesIO, StringIO
+from io import StringIO
 from urllib.parse import urlparse
-from zipfile import ZipFile
 
 import requests
 import requests_cache
 
 from .codelist import Codelist
 from .exceptions import DoesNotExist, NotAvailableInBulk
-from .util import encoding
+from .util import _resolve_zip, encoding
 
 SCHEMAS = ('record-package-schema.json', 'release-package-schema.json', 'release-schema.json')
 
@@ -29,7 +28,6 @@ class ExtensionVersion:
         self.version = data['Version']
         self.base_url = data['Base URL']
         self.download_url = data['Download URL']
-        self._directory = data.get('Directory')
         self._files = None
         self._metadata = None
         self._schemas = None
@@ -50,37 +48,23 @@ class ExtensionVersion:
         for k, v in other.as_dict().items():
             setattr(self, k, v)
 
-    @property
-    def directory(self):
-        return self._directory
-
-    @directory.setter
-    def directory(self, directory):
-        self._directory = directory
-
     def as_dict(self):
         """
         Returns the object's public properties as a dictionary.
         """
         return {key: value for key, value in self.__dict__.items() if not key.startswith('_')}
 
-    def available_in_bulk(self):
-        """
-        Returns whether the files are available in bulk (i.e. whether it has a download URL or a local directory).
-        """
-        return self.directory or self.download_url
-
     def remote(self, basename):
         """
         Returns the contents of the file within the extension.
 
-        If the extension has a download URL or a local directory, caches all the files' contents. Otherwise, downloads
-        and caches the requested file's contents. Raises an HTTPError if a download fails.
+        If the extension has a download URL, caches all the files' contents. Otherwise, downloads and caches the
+        requested file's contents. Raises an HTTPError if a download fails.
 
         :raises DoesNotExist: if the file isn't in the extension
         """
         if basename not in self.files:
-            if not self.available_in_bulk():
+            if not self.download_url:
                 response = requests.get(self.base_url + basename)
                 response.raise_for_status()
                 self._files[basename] = response.text
@@ -95,13 +79,13 @@ class ExtensionVersion:
         """
         Returns the unparsed contents of all files. Decodes the contents of CSV, JSON and Markdown files.
 
-        If the extension has a download URL or a local directory, caches all the files' contents. Otherwise, returns an
-        empty dict. Raises an HTTPError if the download fails.
+        If the extension has a download URL, caches all the files' contents. Otherwise, returns an empty dict. Raises
+        an HTTPError if the download fails.
         """
         if self._files is None:
             self._files = {}
 
-            if self.available_in_bulk():
+            if self.download_url:
                 with closing(self.zipfile()) as zipfile:
                     names = zipfile.namelist()
                     start = len(names[0])
@@ -118,26 +102,14 @@ class ExtensionVersion:
 
     def zipfile(self):
         """
-        If the extension has a local directory, returns a ZIP archive of the files it contains. If the extension has a
-        download URL, downloads and returns the ZIP archive.
+        If the extension has a download URL, downloads and returns the ZIP archive.
 
-        :raises NotAvailableInBulk: if the extension has neither a local directory nor a download URL
+        :raises NotAvailableInBulk: if the extension has no download URL
         """
-        if self.directory:
-            io = BytesIO()
-            with ZipFile(io, 'w') as zipfile:
-                zipfile.write(self.directory)
-                for root, dirs, files in os.walk(self.directory):
-                    for file in sorted(files):
-                        zipfile.write(os.path.join(root, file))
-        elif self.download_url:
-            response = requests.get(self.download_url, allow_redirects=True)
-            response.raise_for_status()
-            io = BytesIO(response.content)
-        else:
-            raise NotAvailableInBulk('ExtensionVersion.zipfile() requires either a directory or a download_url.')
+        if self.download_url:
+            return _resolve_zip(self.download_url)
 
-        return ZipFile(io)
+        raise NotAvailableInBulk('ExtensionVersion.zipfile() requires a download_url.')
 
     @property
     def metadata(self):
@@ -179,7 +151,7 @@ class ExtensionVersion:
 
             if 'schemas' in self.metadata:
                 names = self.metadata['schemas']
-            elif self.available_in_bulk():
+            elif self.download_url:
                 names = [name for name in self.files if name in SCHEMAS]
             else:
                 names = SCHEMAS
@@ -205,7 +177,7 @@ class ExtensionVersion:
 
             if 'codelists' in self.metadata:
                 names = self.metadata['codelists']
-            elif self.available_in_bulk():
+            elif self.download_url:
                 names = [name[10:] for name in self.files if name.startswith('codelists/')]
             else:
                 names = []